Change type of level to short, WIP guild skill database

+++ Scripts/MMOGame/Database/SQLite/SQLiteDatabase_Party.cs
@@ -41,7 +41,7 @@ namespace MultiplayerARPG.MMO
                     partyMemberData.id = reader.GetString("id");
                     partyMemberData.characterName = reader.GetString("characterName");
                     partyMemberData.dataId = reader.GetInt32("dataId");
-                    partyMemberData.level = reader.GetInt32("level");
+                    partyMemberData.level = reader.GetInt16("level");
                     result.AddMember(partyMemberData);
                 }
             }
